% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphviz.R
\name{draw_graphviz_with_images}
\alias{draw_graphviz_with_images}
\title{Draw a graphviz figure that has images in it with grid.}
\usage{
draw_graphviz_with_images(name, package, ...)
}
\description{
Draw a graphviz figure that has images in it with grid.
}
