% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{find_extdata}
\alias{find_extdata}
\title{Return a path to a file included in the specified package.}
\usage{
find_extdata(name, ext, package, inst_dir_name)
}
\description{
Return a path to a file included in the specified package.
}
