% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphviz.R
\name{find_graphviz}
\alias{find_graphviz}
\title{Returns a path to a graphviz file included in the specified package.}
\usage{
find_graphviz(name, package, inst_dir_name = "extdata")
}
\description{
Returns a path to a graphviz file included in the specified package.
}
