% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{find_image}
\alias{find_image}
\title{Returns a path to an image file included in the specified package.}
\usage{
find_image(name, package, inst_dir_name = "extdata")
}
\description{
Returns a path to an image file included in the specified package.
}
