% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{list_extdata}
\alias{list_extdata}
\title{List extdata files available in the specified package by type.}
\usage{
list_extdata(package, re_files, inst_dir_name = "extdata", strip_ext = TRUE)
}
\description{
List extdata files available in the specified package by type.
}

