% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphviz.R
\name{list_graphviz}
\alias{list_graphviz}
\title{List all graphviz files available in the specified package}
\usage{
list_graphviz(package, inst_dir_name = "extdata", strip_ext = TRUE)
}
\description{
List all graphviz files available in the specified package
}

