% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{list_images}
\alias{list_images}
\title{List all graphviz files available in the specified package}
\usage{
list_images(package, strip_ext = TRUE)
}
\description{
List all graphviz files available in the specified package
}
