% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphviz.R
\name{read_all_graphviz_chunks}
\alias{read_all_graphviz_chunks}
\title{Read the graphviz files in a package as knitr chunks.}
\usage{
read_all_graphviz_chunks(package)
}
\description{
Read the graphviz files in a package as knitr chunks.
}

