% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphviz.R
\name{read_graphviz_chunk}
\alias{read_graphviz_chunk}
\title{Read a specific graphviz file as a knitr chunk.}
\usage{
read_graphviz_chunk(name, package, new_name)
}
\arguments{
\item{name}{The name of the graphviz file. Can be the path to
a graphviz file or the basename of a graphviz file
installed in a specific package.}

\item{package}{Optional. The package to search in if the provided
name does not exist.}

\item{new_name}{Optional. Specify the name of the chunk to load.
Defaults to using the basename of the file.}
}
\description{
Read a specific graphviz file as a knitr chunk.
}
