% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphviz.R
\name{read_graphviz_with_images}
\alias{read_graphviz_with_images}
\title{Render a graphviz figure directly with the dot engine.}
\usage{
read_graphviz_with_images(name, package, ...)
}
\description{
Render a graphviz figure directly with the dot engine.
}
