% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphviz.R
\name{read_graphviz_chunk}
\alias{read_graphviz_chunk}
\title{Read a specific graphviz file as a knitr chunk.}
\usage{
read_graphviz_chunk(name, package)
}
\description{
Read a specific graphviz file as a knitr chunk.
}

